scripts/pyproject.toml
[project]
name = "trackstargame-scripts"
version = "0.1.0"
description = "Scripts for scraping Track Star videos and populating database"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "spotipy>=2.23.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]
